+++ packages/sui-lint/bin/sui-lint.js
@@ -4,6 +4,9 @@ const {version} = require('../package.json')
 
 program.version(version, '    --version')
 
-program.command('js', 'lint javascript files').command('sass', 'lint sass files')
+program
+  .command('js', 'lint javascript files')
+  .command('sass', 'lint sass files')
+  .command('repository', 'lint repository structure')
 
 program.parse(process.argv)