+++ packages/sui-ssr/CHANGELOG.md
@@ -1,5 +1,24 @@
 # CHANGELOG
 
+# 8.27.0 (2024-04-05)
+
+
+### Bug Fixes
+
+* copy files ([b692c3e](https://github.com/SUI-Components/sui/commit/b692c3ead33bb81198e15bd69ff01b3a6a5a4fcb))
+* copy ssr dev 404 and 500 html ([27ae05b](https://github.com/SUI-Components/sui/commit/27ae05b0a335f2ca146bc5d46d2f39a394d54234))
+
+
+
+# 8.26.0 (2024-04-05)
+
+
+### Bug Fixes
+
+* internal error ([c7e2524](https://github.com/SUI-Components/sui/commit/c7e252499ada037c4c2d887863378cde192e1c24))
+
+
+
 # 8.25.0 (2024-01-11)
 
 
@@ -1535,6 +1554,3 @@
 ### Features
 
 * **sui-ssr:** add package.json ([d1495ce](https://github.com/SUI-Components/sui/commit/d1495ce76b02d906ca512f59518737411a67acc1)), closes [#67](https://github.com/SUI-Components/sui/issues/67)
-
-
-